+++ index.php
@@ -1,22 +1,45 @@
 <?php
 require_once "inc/functions.php";
 
+session_name('auth');
+session_start();
+
 $task = $_GET['task'] ?? 'report';
 $error = $_GET['error'] ?? '0';
-if ('seed' == $task) {
-    seed();
-    $error = 2;
-}
-if ('delete' == $task) {
-    $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
-    if ($id > 0) {
-        deleteStudent($id);
+
+// Edit
+if ('edit' == $task) {
+    if (!hasPrivilege()) {
         header('location: /');
     }
 }
+
+
+// Add Dummy Students
+if ('seed' == $task) {
+    if (isRole('admin')) {
+        seed();
+        $error = 2;
+    }
+}
+
+// Delete Student from List
+if ('delete' == $task) {
+    if (isRole('admin')) {
+        $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
+        if ($id > 0) {
+            deleteStudent($id);
+            header('location: /');
+        }
+    } else {
+        header('location: /');
+    }
+}
+
 $fname = '';
 $lname = '';
 $roll = '';
+
 if (isset($_POST['submit'])) {
     $fname = filter_input(INPUT_POST, 'fname', FILTER_SANITIZE_STRING);
     $lname = filter_input(INPUT_POST, 'lname', FILTER_SANITIZE_STRING);
@@ -91,7 +114,7 @@ if (isset($_POST['submit'])) {
     </div>
     <?php endif;?>
 
-    <?php if ('add' == $task): ?>
+    <?php if ('add' == $task && hasPrivilege()): ?>
     <div class="container">
         <div class="row">
             <div class="column column-60 column-offset-20">
@@ -111,30 +134,30 @@ if (isset($_POST['submit'])) {
 
 
     <?php
-if ('edit' == $task):
-    $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
-    $student = getStudent($id);
-    if ($student): ?>
-	            <div class="container">
-	                <div class="row">
-	                    <div class="column column-60 column-offset-20">
-	                        <form action="/?task=add" method="POST">
-	                            <input type="hidden" value="<?php echo $id; ?>" name="id">
-	                            <label for="fname">First Name</label>
-	                            <input type="text" name="fname" value="<?php echo $student['fname']; ?>">
-	                            <label for="lname">Last Name</label>
-	                            <input type="text" name="lname" value="<?php echo $student['lname']; ?>">
-	                            <label for="roll">Roll</label>
-	                            <input type="number" name="roll" value="<?php echo $student['roll']; ?>">
-	                            <button type="submit" name="submit"> Update </button>
-	                        </form>
-	                    </div>
-	                </div>
-	            </div>
-	        <?php
-endif;
-endif;
-?>
+    if ('edit' == $task && hasPrivilege()):
+        $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
+        $student = getStudent($id);
+        if ($student): ?>
+        <div class="container">
+            <div class="row">
+                <div class="column column-60 column-offset-20">
+                    <form action="/?task=add" method="POST">
+                        <input type="hidden" value="<?php echo $id; ?>" name="id">
+                        <label for="fname">First Name</label>
+                        <input type="text" name="fname" value="<?php echo $student['fname']; ?>">
+                        <label for="lname">Last Name</label>
+                        <input type="text" name="lname" value="<?php echo $student['lname']; ?>">
+                        <label for="roll">Roll</label>
+                        <input type="number" name="roll" value="<?php echo $student['roll']; ?>">
+                        <button type="submit" name="submit"> Update </button>
+                    </form>
+                </div>
+            </div>
+        </div>
+        <?php
+        endif;
+    endif;
+    ?>
     <script type="text/javascript" src="assets/js/main.js"></script>
 </body>
 </html>
